refactor: replace fetchDataFromDGS with useDGSData hook for improved data handling in DGSKeyStatistics component
